+++ src/main/java/com/iemr/mmu/repo/login/UserParkingplaceMappingRepo.java
@@ -31,7 +31,7 @@ import org.springframework.stereotype.Repository;
 import com.iemr.mmu.data.login.UserParkingplaceMapping;
 
 @Repository
-public interface UserParkingplaceMappingRepo extends CrudRepository<UserParkingplaceMapping, Long> {
+public interface UserParkingplaceMappingRepo extends CrudRepository<UserParkingplaceMapping, Integer> {
 	@Query("SELECT x.parkingPlaceID,p.stateID,s.stateName,p.districtID,d.districtName,p.districtBlockID,b.blockName from UserParkingplaceMapping x "
 			+ " INNER JOIN x.m_parkingplace p"
 			+ " INNER JOIN p.state s"
