+++ src/main/java/com/iemr/mmu/repo/masterrepo/doctor/PreMalignantLesionMasterRepo.java
@@ -29,7 +29,7 @@ import org.springframework.stereotype.Repository;
 
 import com.iemr.mmu.data.masterdata.doctor.PreMalignantLesion;
 @Repository
-public interface PreMalignantLesionMasterRepo extends CrudRepository<PreMalignantLesion, Long>{
+public interface PreMalignantLesionMasterRepo extends CrudRepository<PreMalignantLesion, Integer>{
 
 	@Query("select preMalignantLesionID, preMalignantLesionType from PreMalignantLesion where deleted = false "
 			+ "order by preMalignantLesionType")
